requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pymongo==4.6.0
bcrypt==4.1.2